Fixed two issues: (1) port, (2) ROCM test failures in CI
