+++ test/common.py
@@ -17,6 +17,7 @@ import unittest
 import warnings
 import random
 import contextlib
+import socket
 from functools import wraps
 from itertools import product
 from copy import deepcopy
@@ -550,3 +551,12 @@ def download_file(url, binary=True):
         msg = "could not download test file '{}'".format(url)
         warnings.warn(msg, RuntimeWarning)
         raise unittest.SkipTest(msg)
+
+
+def find_free_port():
+    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+    sock.bind(('localhost', 0))
+    sockname = sock.getsockname()
+    sock.close()
+    return sockname[1]
